Refactoring: decoupled tasks from component
